requirements.txt
Django==2.0.12
pycryptodome==3.6.6
djangorestframework==3.9.1
django-cors-headers==2.1.0
requests==2.20.0
django-filter==1.1.0
psycopg2-binary==2.8.6
django-rest-swagger==2.2.0
coverage==4.5.2
django-nose==1.4.6
jsonnet==0.12.1
selenium
click
locust
python-social-auth
social-auth-app-django
pandas
xlrd
openpyxl
python-decouple==3.6